# backend/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/time_template": {
            "get": {
                "tags": [
                    "time_template"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page index",
                        "name": "page_index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "template category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "filter by active status",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "info mode",
                        "name": "info",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TimeTemplate"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "time_template"
                ],
                "parameters": [
                    {
                        "description": "time template",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TimeTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
        "/time_template/:id": {
            "put": {
                "tags": [
                    "time_template"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "template id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "time template",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TimeTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "time_template"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "template id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
...
        "/time_template/check": {
            "post": {
                "tags": [
                    "time_template"
                ],
                "parameters": [
                    {
                        "description": "time access check request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CheckTimeAccessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.CheckTimeAccessResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.CheckTimeAccessRequest": {
            "type": "object",
            "required": [
                "template_id"
            ],
            "properties": {
                "template_id": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "controller.CheckTimeAccessResponse": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "checked_at": {
                    "type": "string"
                },
                "template_id": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "controller.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.TimeRange": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "weekdays": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "model.TimeTemplate": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "work, maintenance, emergency, etc.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_builtin": {
                    "description": "Status and metadata",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "resource_id": {
                    "description": "Standard fields",
                    "type": "integer"
                },
                "time_ranges": {
                    "description": "Time configuration",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TimeRange"
                    }
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updater_id": {
                    "type": "integer"
                },
                "usage_count": {
                    "description": "Usage statistics",
                    "type": "integer"
                }
            }
        },
...
    }
}